+++ routes/builderRoutes/list.js
@@ -5,16 +5,33 @@ const firebase = require('../../firebase.js').builder
 const SeriesReal = require('../../models/builderModels/models-real/Series.js')
 const SeriesOther = require('../../models/builderModels/models-other/Series.js')
 const LessonOther = require('../../models/builderModels/models-other/Lesson.js')
+const SermonOther = require('../../models/builderModels/models-other/Sermon.js')
 
 const Topic = require('../../models/nqModels/Topic.js')
 
-var realContent = {
+// Import other media types
+const Quote = require('../../models/builderModels/models-other/Quote.js')
+const Image = require('../../models/builderModels/models-other/Image.js')
+const Illustration = require('../../models/builderModels/models-other/Illustration.js')
+const Lyric = require('../../models/builderModels/models-other/Lyric.js')
+const Video = require('../../models/builderModels/models-other/Video.js')
+
+const realContent = {
     'rseries': SeriesReal
 }
 
-var otherContent = {
+const otherContent = {
     'oseries': SeriesOther,
-    'olessons': LessonOther
+    'olessons': LessonOther,
+    'osermons': SermonOther
+}
+
+const newMedia = {
+    'quote': Quote,
+    'image': Image,
+    'illustration': Illustration,
+    'lyric': Lyric,
+    'video': Video
 }
 
 module.exports = function (req, res) {
@@ -39,6 +56,12 @@ module.exports = function (req, res) {
                 })
             } else if (type === 'topic') {
                 Topic.find({}).exec(function (err, items) {
+                    if (err) console.log(err)
+                    console.log(items)
+                    res.send(items)
+                })
+            } else if (Object.keys(newMedia).includes(type)) {
+                newMedia[type].find({ user: decodedToken.uid }).exec(function (err, items) {
                     if (err) console.log(err)
                     console.log(items)
                     res.send(items)